feat: add challenge one and two resolutions
Challenge one is a algorithm that prints the whole integer numbers to the console, start from the
number 1, and print all numbers going up to the number 100. However, when the number is divisible by
3, do not print the number, but print the word 'Visual'. If the number is divisible by 5, do not
print the number, but print 'Nuts'. And for all numbers divisible by both (eg: the number 15) the
same, but print 'Visual Nuts'

Challenge two is a service that shows information about official languages spoken by countries